updated .gitignore

+++ app/controllers/activities_controller.rb
@@ -1,4 +1,5 @@
 class ActivitiesController < ApplicationController
+  before_action :authenticate
   before_action :set_activity, only: [:show, :edit, :update, :destroy]
 
   # GET /activities
@@ -51,11 +52,18 @@ class ActivitiesController < ApplicationController
   # PATCH/PUT /activities/1
   # PATCH/PUT /activities/1.json
   def update
-    respond_to do |format|
-      if @activity.update(activity_params)
-        format.json { render :json => {:success=>true, :message=>"Activity was successfully updated."} }
-      else
-        format.json { render :json => {:success=>true, :message=>"Something went wrong. Could not update activity."} }
+    # only the owner can edit his / her activity
+    if current_user.id == @activity.user_id
+      respond_to do |format|
+        if @activity.update(activity_params)
+          format.json { render :json => {:success=>true, :message=>"Activity was successfully updated."} }
+        else
+          format.json { render :json => {:success=>true, :message=>"Something went wrong. Could not update activity."} }
+        end
+      end
+    else
+      respond_to do |format|
+        format.json { render :json => {:success=>false, :message=>"You are not the owner of this activity."} }
       end
     end
   end
@@ -63,14 +71,21 @@ class ActivitiesController < ApplicationController
   # DELETE /activities/1
   # DELETE /activities/1.json
   def destroy
-    # also delete all goingtoactivities that have this actvity
-    Goingtoactivity.where(activity_id: @activity.id).delete_all
+    # only the owner can destroy his / her activity
+    if current_user.id == @activity.user_id
+      # also delete all goingtoactivities that have this actvity
+      Goingtoactivity.where(activity_id: @activity.id).delete_all
 
-    @activity.destroy
-    respond_to do |format|
-      format.json { render :json => {:success=>true, :message=>"Activity was successfully destroyed."} }
+      @activity.destroy
+      respond_to do |format|
+        format.json { render :json => {:success=>true, :message=>"Activity was successfully destroyed."} }
+      end
+      # TODO: what if the activity wasn't successfully destroyed?
+    else
+      respond_to do |format|
+        format.json { render :json => {:success=>false, :message=>"You are not the owner of this activity."} }
+      end
     end
-    # TODO: what if the activity wasn't successfully destroyed?
   end
 
   private
